Close semantic type system and validate checked shape extents
The semantic type system conflated type, shape, and resource
representations across open enums and host usize fields.

Remove non_exhaustive attributes from DataType, BufferAccess, and
ShapeIndexRelation to make the semantic type system closed and
exhaustive. Replace host usize with fixed-width u32 and u64 types in
DataType::Array and BufferDecl ranges. Distinguish zero-extent shapes
from unresolved dimensions with ShapeDim::Unresolved and
ShapeDim::Expr. Enforce checked arithmetic across Tile and layout
computations, and bump the wire format version to reject stale records.

+++ vyre-reference/src/value.rs
@@ -350,7 +350,7 @@ fn fixed_scalar_storage_width(ty: &vyre_foundation::ir::DataType) -> Option<usiz
         vyre_foundation::ir::DataType::Handle(_)
         | vyre_foundation::ir::DataType::DeviceMesh { .. } => Some(4),
         vyre_foundation::ir::DataType::I64 => Some(8),
-        vyre_foundation::ir::DataType::Array { element_size } => Some(*element_size),
+        vyre_foundation::ir::DataType::Array { element_size } => Some(*element_size as usize),
         vyre_foundation::ir::DataType::Vec { element, count } => {
             fixed_scalar_storage_width(element)
                 .and_then(|width| width.checked_mul(usize::from(*count)))
